improved curl downloader + try all URLs
We now try all URLs when remote fetching,
unless the user has explicitly chosen a single station

We centralise all curl calls, which solves problem previously with redirects being seen as
successes and storing the redirect page
instead of the almanac.

ephemeris data is now stored in the file passed
by the command line if given. And then the same
file is read for current ephemeris data

We lengthen the timeout on starting the simulator thread since the fetch calls could take a little while in pathological cases